feat: production hardening — TLS/mTLS, timeout/deadline interceptors, weighted round-robin balancer, CORS config, readiness health, CI pipeline, queue reclaim, go-redis extraction, and full sample compatibility pass
Core:
- server: register health service, wire timeout/recovery interceptors, TLS/mTLS support, net.SplitHostPort, restart safety
- app: initConfig idempotent in beforeRun, restart re-registers services+registry, health dependency tracking
- naming/etcd: lease KeepAlive with exponential backoff re-registration, weighted round-robin balancer, List stub, Connection Close
- client: connection pooling/caching, retryPolicy, healthCheckConfig, keepalive/connect params, reasonable msg/window sizes
- broker/nats: non-blocking subscribe, reconnect/disconnect handlers, consistent topic prefix, broker.Subscription interface

Features:
- interceptor/timeout: server default timeout (60s) + client deadline validation
- internal/tls: mTLS server/client transport credentials
- naming/etcd/balancer: weighted_round_robin_xds with balancer attribute weights

Reliability:
- queue: atomic ClaimJob + processing set lifecycle (Pop/Update cleanup), ReclaimExpired with 10s ticker, FetchJob window fix, pipeline Push/Clear, graceful Stop with context
- cache/redis: real implementations (IsExist/Incr/Decr/Fetch/ClearAll), Put default semantics, ClearAll prefix safety
- store: InstanceE error propagation, mysql DSN timezone config, uid uint64+Close+tx timeout
- interceptor/ratelimit: limiterEntry cleanup goroutine, Stop()
- interceptor/recovery: Stream panic recovery
- api: http.Server with timeouts, graceful shutdown, NewApi returns error, CORS config, TLS

Engineering:
- CI: .github/workflows/ci-build.yml with -race, per-module/per-sample matrix build+vet
- store/postgres: InstanceE implementation
- store/redis/limiter: deprecated non-atomic methods
- store/redis: GetCluster TODO comment, naming/etcd List warn
- config/etcd: Watch handles Delete/put, reconnects, Close()
- config: Validate enforced at parse time
- AGENTS.md: updated Go version to 1.26.4
- interceptor/ratelimit/redis: extracted go-redis from root into nested module

Samples: all go.mod/go.sum synced, gateway/api/cache/handler adapted to new signatures

+++ store/mysql/uid/options.go
@@ -5,8 +5,8 @@ import (
 	"gorm.io/gorm"
 )
 
-type CheckFunc func(num ...uint32) uint32
-type SetpFunc func(num uint32) uint32
+type CheckFunc func(num ...uint64) uint64
+type SetpFunc func(num uint64) uint64
 
 type Options struct {
 	table  string //表名称
